订单自动取消

+++ shop-service-order/shop-service-order-api/src/main/java/com/shop/order/OrderApi.java
@@ -12,5 +12,5 @@ public interface OrderApi {
     String PREFIX = ApiConstants.PREFIX + "/order";
 
     @PostMapping( PREFIX + "/get")
-    OrderRespDTO getOrderById(@RequestBody Long id);
+    OrderRespDTO getOrderById(@RequestBody Integer id);
 }
